Enable the use of dynamic termination detection in PTG via a parsec-ptgpp flag (work in progress)

+++ parsec/mca/termdet/user_trigger/termdet_user_trigger.h
@@ -39,7 +39,7 @@ PARSEC_DECLSPEC extern const parsec_termdet_module_t parsec_termdet_user_trigger
 /* static accessor */
 mca_base_component_t *termdet_user_trigger_static_component(void);
 
-int parsec_termdet_user_trigger_msg_dispatch(parsec_comm_engine_t *ce, unsigned long long tag,  void *msg,
+int parsec_termdet_user_trigger_msg_dispatch(parsec_comm_engine_t *ce, unsigned long tag,  void *msg,
                                              unsigned long size, int src,  void *module);
 
 typedef struct {
